summary: hook for better initial starting value
When the service starts, the summary is automatically set to the
value of the first starting unit. This allows overriding the
starting value for some summaries, while still using the value
from the first unit as a hint.

Later we will use that to prevent settings-based summaries from
initialising at startup, and instead delay this until later.