Improve AC current limit setting
Instead of fetching the min and max and trying to work with that,
simply set the current limit (if adjustable) and let the
ItemsChanged coming back from the units update the path. That way
it will always update to the maximum the units support implicitly
without having to try doing this ourselves with the min and max
values.